타워 버그수정 및 데미지 설정 가능하게 변경
타워가 포탄의 데미지를 가집니다

+++ Assets/Scripts/InGame/GameObject/Enemy/EnemyDamage.cs
@@ -13,8 +13,8 @@ public class EnemyDamage : MonoBehaviour
     private Canvas uiCanvas;
     public Image hpBarImage;
 
-    public int CurHp;
-    public int InitHp;
+    public float CurHp;
+    public float InitHp;
 
     public bool isDie;
